General UI improvements.
Imported the diffs from http://code.google.com/p/munki/issues/detail?id=194 and
http://code.google.com/p/munki/issues/detail?id=193 which allow a username to be
added to a manifest where the username is initially blank and cleans up the
description and *_script fields in the catalogs view.

Also added a little javascript to the catalog and manifest pages that resizes
the repesctive tables so they fit within the browser. This includes a live
resize for when the user changes the size of the browser.