Fixed a positioning error with the manifests table.
If the "Delete manifest" control was present when the "New manifest" control was
not, the manifests table was pushing the "Delete manifest" control off the
bottom of the page. It now checks for the height of the "New manifest" button
and falls back to the height of the "Delete manifest" button (+30px in both
cases) when calculating the appropriate height of the manifests table.