Minor cleanup of Adam Reed's changes.
Changed WARRANTY to WARRANY_LOOKUP_ENABLED in settings_template.py just to be
verbose.

Moved the Report and Inventory links to the subtext of the machine hostname -
makes it feel a bit more like tabs, but more importantly it puts those
navigation links in a consistent location. Moved the last checkin date up there
as well since that seems like information I find myself looking for frequently.

Moved Warranty information to its own table below the Munki table on the machine
reports page. The "Machine" table is already tall enough and adding the warranty
information there made it seem awkwardly taller than the right side of the page.
Also renamed "Estimated manufacture date" to "Manufacture date" to help prevent
text wrapping; also, the values are "Week of ...", so it's pretty obvious that
it's an estimated date.